+++ lib/ui/profile/ProfileInfoPage.dart
@@ -8,7 +8,6 @@ import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/ui/directory/DirectoryAccountsPage.dart';
 import 'package:illinois/ui/profile/ProfileInfoEditPage.dart';
 import 'package:illinois/ui/profile/ProfileInfoPreviewPage.dart';
-import 'package:illinois/ui/profile/ProfileInfoAndDirectoryPage.dart';
 import 'package:illinois/ui/directory/DirectoryWidgets.dart';
 import 'package:illinois/ui/profile/ProfileLoginPage.dart';
 import 'package:illinois/ui/settings/SettingsWidgets.dart';
@@ -25,15 +24,18 @@ import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
+enum ProfileInfo { connectionsInfo, directoryInfo }
+
 class ProfileInfoPage extends StatefulWidget {
   static const String editParamKey = 'edu.illinois.rokwire.profile.directory.info.edit';
 
   final ProfileInfo contentType;
   final Map<String, dynamic>? params;
+  final bool showProfileCommands;
   final bool showAccountCommands;
   final void Function()? onStateChanged;
 
-  ProfileInfoPage({super.key, required this.contentType, this.params, this.showAccountCommands = false, this.onStateChanged});
+  ProfileInfoPage({super.key, required this.contentType, this.params, this.showProfileCommands = true, this.showAccountCommands = false, this.onStateChanged});
 
   @override
   State<StatefulWidget> createState() => ProfileInfoPageState();
@@ -46,6 +48,9 @@ class ProfileInfoPage extends StatefulWidget {
 
 class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileInfoPage> implements NotificationsListener {
 
+  final GlobalKey<ProfileInfoPreviewPageState> _profileInfoPreviewKey = GlobalKey<ProfileInfoPreviewPageState>();
+  final GlobalKey<ProfileInfoEditPageState> _profileInfoEditKey = GlobalKey<ProfileInfoEditPageState>();
+
   Auth2UserProfile? _profile;
   Auth2UserPrivacy? _privacy;
   Uint8List? _photoImageData;
@@ -57,15 +62,28 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
   bool _updatingDirectoryVisibility = false;
   bool _preparingDeleteAccount = false;
 
-  bool get previewMode => (!_loading && !_editing);
-  bool get _directoryVisibility => (_privacy?.public == true);
+  bool get isEditing => _editing;
+  bool get isLoading => _loading;
+  bool get directoryVisibility => (_privacy?.public == true);
+
+  void setEditing(bool value) {
+    if (mounted && (_editing != value)) {
+      setState(() {
+        _editing = value;
+        widget.onStateChanged?.call();
+      });
+    }
+  }
+
+  Future<bool> saveEdit() =>
+    _profileInfoEditKey.currentState?.saveEdit() ?? Future.value(false);
 
   @override
   void initState() {
     NotificationService().subscribe(this, [
       DirectoryAccountsPage.notifyEditInfo,
     ]);
-    _editing = widget.editParam ?? false;
+    _editing = (widget.editParam == true);
     _loadInitialContent();
     super.initState();
   }
@@ -95,7 +113,7 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
       return Column(children: [
         _directoryVisibilityControl,
 
-        if (_directoryVisibility == true)
+        if (directoryVisibility == true)
           Column(children: [
             Padding(padding: EdgeInsets.symmetric(vertical: 16), child:
               Text(_desriptionText, style: Styles().textStyles.getTextStyle('widget.detail.small'), textAlign: TextAlign.center,),
@@ -114,6 +132,7 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
     Padding(padding: EdgeInsets.symmetric(horizontal: 16), child:
       Column(children: [
         ProfileInfoPreviewPage(
+          key: _profileInfoPreviewKey,
           contentType: widget.contentType,
           profile: _profile,
           privacy: _privacy,
@@ -121,20 +140,23 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
           photoImageData: _photoImageData,
           photoImageToken: _photoImageToken,
         ),
-        Padding(padding: EdgeInsets.only(top: 24), child:
-          _previewCommandBar,
-        ),
+        if (widget.showProfileCommands)
+          Padding(padding: EdgeInsets.only(top: 24), child:
+            _previewCommandBar,
+          ),
       ]),
     );
 
   Widget get _editContent =>
     ProfileInfoEditPage(
+      key: _profileInfoEditKey,
       contentType: widget.contentType,
       profile: _profile,
       privacy: _privacy,
       pronunciationAudioData: _pronunciationAudioData,
       photoImageData: _photoImageData,
       photoImageToken: _photoImageToken,
+      showProfileCommands: widget.showProfileCommands,
       onFinishEdit: _onFinishEditInfo,
   );
 
@@ -162,7 +184,7 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
   Widget get _directoryVisibilityToggleButton =>
     InkWell(onTap: _onToggleDirectoryVisibility, child:
       Padding(padding: EdgeInsets.only(left: 16, right: 16, top: 16, bottom: 12), child:
-        Styles().images.getImage(_directoryVisibility ? 'toggle-on' : 'toggle-off')
+        Styles().images.getImage(directoryVisibility ? 'toggle-on' : 'toggle-off')
       )
     );
 
@@ -178,11 +200,11 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
   Widget get _directoryVisibilityDescription {
     final String visibilityMacro = "{{visibility}}";
 
-    final String visibilityValue = _directoryVisibility ?
+    final String visibilityValue = directoryVisibility ?
       Localization().getStringEx('panel.profile.info.directory_visibility.public.text', 'Public') :
       Localization().getStringEx('panel.profile.info.directory_visibility.private.text', 'Private');
 
-    final String messageTemplate = _directoryVisibility ?
+    final String messageTemplate = directoryVisibility ?
       Localization().getStringEx('panel.profile.info.directory_visibility.public.description', 'Your directory visibility is set to $visibilityMacro. Anyone on or off the User Directory can view your account.') :
       Localization().getStringEx('panel.profile.info.directory_visibility.private.description', 'Your directory visibility is set to $visibilityMacro. Your account is available only to you.');
 
@@ -201,12 +223,12 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
   }
 
   void _onToggleDirectoryVisibility() {
-    Analytics().logSelect(target: "Directory Visibility: ${_directoryVisibility ? 'OFF' : 'ON'}");
+    Analytics().logSelect(target: "Directory Visibility: ${directoryVisibility ? 'OFF' : 'ON'}");
     setState(() {
       _updatingDirectoryVisibility = true;
     });
     Auth2UserPrivacy privacy = Auth2UserPrivacy.fromOther(_privacy,
-      public: !_directoryVisibility,
+      public: !directoryVisibility,
     );
     Auth2().saveUserPrivacy(privacy).then((bool result){
       if (mounted) {
@@ -262,7 +284,7 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
     Expanded(child: _editInfoButton,),
     Container(width: 8),
     Expanded(child: _swapInfoButton,),
-  ],);
+  ]);
 
   Widget get _myDirectoryInfoPreviewCommandBar => Row(children: [
     Expanded(flex: 1, child: Container(),),
@@ -487,7 +509,9 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
         _pronunciationAudioData = pronunciationAudioData;
       }
 
-      _editing = false;
+      if (widget.showProfileCommands) {
+        _editing = false;
+      }
       widget.onStateChanged?.call();
     });
   }