+++ CHANGELOG.md
@@ -6,10 +6,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
-- Acknowledged global message id parameter from social message notification [#4572](https://github.com/rokwire/illinois-app/issues/4572).
 - Edit a message [#4579](https://github.com/rokwire/illinois-app/issues/4572)
+
+## [6.1.37] - 2025-01-10
+### Changed
+- UI Cleanup: Home > Sections > news and radio groupings [#4469](https://github.com/rokwire/illinois-app/issues/4469).
+- Rename "New Message" button to "New Connection" on the Connections panel [#4594](https://github.com/rokwire/illinois-app/issues/4594).
+- GroupDetailPanel UI updates (in progress) [#4483](https://github.com/rokwire/illinois-app/issues/4483).
+- Updated Onboarding Profile Info flow [#4596](https://github.com/rokwire/illinois-app/issues/4596).
+
+## [6.1.36] - 2025-01-09
+### Added
+- Acknowledged global message id parameter from social message notification [#4572](https://github.com/rokwire/illinois-app/issues/4572).
+- Acknowledged "extend-limit-to-global-message-id" parameter of GET /api/conversations/{id}/messages API [#4572](https://github.com/rokwire/illinois-app/issues/4572).
 ### Fixed
 - Fixed pagination in conversation messages panel [#4572](https://github.com/rokwire/illinois-app/issues/4572).
+### Changed
+- GroupDetailPanel UI updates (in progress) [#4483](https://github.com/rokwire/illinois-app/issues/4483).
+- "Messages" renamed to "Connections" [#4585](https://github.com/rokwire/illinois-app/issues/4585).
+- Change the name in sections from "User Directory" to "Directory of Users" [#4587](https://github.com/rokwire/illinois-app/issues/4587).
+- Old Profile page replaced by stripped version of Profile Info and Connections page [#4589](https://github.com/rokwire/illinois-app/issues/4589).
+### Added
+- Added Share My Info sheet [#4589](https://github.com/rokwire/illinois-app/issues/4589).
  
 ## [6.1.35] - 2025-01-08
 ### Changed